fix(avform): Audio/Video settings tab #5011 issue bugfix

+++ src/widget/form/settings/avform.cpp
@@ -153,7 +153,7 @@ void AVForm::open(const QString& devName, const VideoMode& mode)
 {
     QRect rect = mode.toRect();
     videoSettings->setCamVideoRes(rect);
-    videoSettings->setCamVideoFPS(static_cast<quint16>(mode.FPS));
+    videoSettings->setCamVideoFPS(static_cast<float>(mode.FPS));
     camera.setupDevice(devName, mode);
 }
